Rename LocationUpdateTask-> LocationEstimateTask Define a New UIUpdateTask: UIUpdateTask_map to update the Map UI. UIUpdateTask_Data will no longer be used in this phase. Upload Jessup.png draw by Harpreet.

+++ WifiLocator/src/wifilocator/gui/LocationDraw.java
@@ -25,7 +25,7 @@ public class LocationDraw {
 		canvas=new Canvas(map);
 	}
 	
-	public void draw(int x,int y)
+	public void draw(float x,float y)
 	{
 		canvas.save();
 		Paint paint = new Paint();
